# Make head Postings only return series in time range