--- a/tsdb/cmd/tsdb/main.go
+++ b/tsdb/cmd/tsdb/main.go
@@ -470,7 +470,7 @@ func analyzeBlock(b tsdb.BlockReader, limit int) error {
 	// Presume 1ms resolution that Prometheus uses.
 	fmt.Printf("Duration: %s\n", (time.Duration(meta.MaxTime-meta.MinTime) * 1e6).String())
 	fmt.Printf("Series: %d\n", meta.Stats.NumSeries)
-	ir, err := b.Index()
+	ir, err := b.Index(math.MinInt64, math.MaxInt64)
 	if err != nil {
 		return err
 	}