--- a/tsdb/head.go
+++ b/tsdb/head.go
@@ -754,7 +754,7 @@ type RangeHead struct {
 	mint, maxt int64
 }
 
-// NewRangeHead returns a *rangeHead.
+// NewRangeHead returns a *RangeHead.
 func NewRangeHead(head *Head, mint, maxt int64) *RangeHead {
 	return &RangeHead{
 		head: head,
@@ -764,7 +764,14 @@ func NewRangeHead(head *Head, mint, maxt int64) *RangeHead {
 }
 
 func (h *RangeHead) Index(mint, maxt int64) (IndexReader, error) {
-	return h.head.indexRange(h.mint, h.maxt), nil
+	// rangeHead guarantees that the series returned are within its range.
+	if mint < h.mint {
+		mint = h.mint
+	}
+	if maxt > h.maxt {
+		maxt = h.maxt
+	}
+	return h.head.indexRange(mint, maxt), nil
 }
 
 func (h *RangeHead) Chunks() (ChunkReader, error) {
@@ -1347,9 +1354,17 @@ func (h *headIndexReader) LabelNames() ([]string, error) {
 
 // Postings returns the postings list iterator for the label pairs.
 func (h *headIndexReader) Postings(name string, values ...string) (index.Postings, error) {
+	fullRange := h.mint <= h.head.MinTime() && h.maxt >= h.head.MaxTime()
 	res := make([]index.Postings, 0, len(values))
 	for _, value := range values {
 		p := h.head.postings.Get(name, value)
+		if fullRange {
+			// The head timerange covers the full index reader timerange.
+			// All the series can the be appended without filtering.
+			res = append(res, p)
+			continue
+		}
+
 		// Filter out series not in the time range, to avoid
 		// later on building up all the chunk metadata just to
 		// discard it.