--- a/tsdb/querier_bench_test.go
+++ b/tsdb/querier_bench_test.go
@@ -16,6 +16,7 @@ package tsdb
 import (
 	"fmt"
 	"io/ioutil"
+	"math"
 	"os"
 	"strconv"
 	"testing"
@@ -54,7 +55,7 @@ func BenchmarkPostingsForMatchers(b *testing.B) {
 	}
 	testutil.Ok(b, app.Commit())
 
-	ir, err := h.Index()
+	ir, err := h.Index(math.MinInt64, math.MaxInt64)
 	testutil.Ok(b, err)
 	b.Run("Head", func(b *testing.B) {
 		benchmarkPostingsForMatchers(b, ir)
@@ -72,7 +73,7 @@ func BenchmarkPostingsForMatchers(b *testing.B) {
 	defer func() {
 		testutil.Ok(b, block.Close())
 	}()
-	ir, err = block.Index()
+	ir, err = block.Index(math.MinInt64, math.MaxInt64)
 	testutil.Ok(b, err)
 	defer ir.Close()
 	b.Run("Block", func(b *testing.B) {