@roidelapluie 2020-01-21
benchmark                                  old ns/op      new ns/op      delta
BenchmarkHeadSeries/1of1000000-8           1182394114     111903957      -90.54%
BenchmarkHeadSeries/10of1000000-8          1161618692     111403226      -90.41%
BenchmarkHeadSeries/100of1000000-8         1169758844     112122961      -90.41%
BenchmarkHeadSeries/1000of1000000-8        1175302734     113658484      -90.33%
BenchmarkHeadSeries/10000of1000000-8       1191381454     123592113      -89.63%
BenchmarkHeadSeries/100000of1000000-8      1186973443     237907048      -79.96%
BenchmarkHeadSeries/1000000of1000000-8     1358181144     1374416302     +1.20%

benchmark                                  old allocs     new allocs     delta
BenchmarkHeadSeries/1of1000000-8           4000055        28             -100.00%
BenchmarkHeadSeries/10of1000000-8          4000073        87             -100.00%
BenchmarkHeadSeries/100of1000000-8         4000253        630            -99.98%
BenchmarkHeadSeries/1000of1000000-8        4002053        6036           -99.85%
BenchmarkHeadSeries/10000of1000000-8       4020053        60054          -98.51%
BenchmarkHeadSeries/100000of1000000-8      4200053        600074         -85.71%
BenchmarkHeadSeries/1000000of1000000-8     6000053        6000053        +0.00%

benchmark                                  old bytes     new bytes     delta
BenchmarkHeadSeries/1of1000000-8           229192185     2488          -100.00%
BenchmarkHeadSeries/10of1000000-8          229193337     5568          -100.00%
BenchmarkHeadSeries/100of1000000-8         229204857     35536         -99.98%
BenchmarkHeadSeries/1000of1000000-8        229320271     345104        -99.85%
BenchmarkHeadSeries/10000of1000000-8       230472057     3894672       -98.31%
BenchmarkHeadSeries/100000of1000000-8      241992057     40511632      -83.26%
BenchmarkHeadSeries/1000000of1000000-8     357192058     357192058     +0.00%