@roidelapluie 2020-01-21
```
BenchmarkHeadSeries/1of1000000-8           1182394114     111477804      -90.57%
BenchmarkHeadSeries/10of1000000-8          1161618692     111178736      -90.43%
BenchmarkHeadSeries/100of1000000-8         1169758844     112278785      -90.40%
BenchmarkHeadSeries/1000of1000000-8        1175302734     112791163      -90.40%
BenchmarkHeadSeries/10000of1000000-8       1191381454     122281323      -89.74%
BenchmarkHeadSeries/100000of1000000-8      1186973443     240413878      -79.75%
BenchmarkHeadSeries/1000000of1000000-8     1358181144     1387995740     +2.20%

benchmark                                  old allocs     new allocs     delta
BenchmarkHeadSeries/1of1000000-8           4000055        28             -100.00%
BenchmarkHeadSeries/10of1000000-8          4000073        87             -100.00%
BenchmarkHeadSeries/100of1000000-8         4000253        630            -99.98%
BenchmarkHeadSeries/1000of1000000-8        4002053        6036           -99.85%
BenchmarkHeadSeries/10000of1000000-8       4020053        60054          -98.51%
BenchmarkHeadSeries/100000of1000000-8      4200053        600074         -85.71%
BenchmarkHeadSeries/1000000of1000000-8     6000053        6000053        +0.00%

benchmark                                  old bytes     new bytes     delta
BenchmarkHeadSeries/1of1000000-8           229192185     2488          -100.00%
BenchmarkHeadSeries/10of1000000-8          229193337     5568          -100.00%
BenchmarkHeadSeries/100of1000000-8         229204857     35536         -99.98%
BenchmarkHeadSeries/1000of1000000-8        229320271     345104        -99.85%
BenchmarkHeadSeries/10000of1000000-8       230472057     3894672       -98.31%
BenchmarkHeadSeries/100000of1000000-8      241992057     40511632      -83.26%
BenchmarkHeadSeries/1000000of1000000-8     357192058     357192058     +0.00%
```