@brian-brazil 2020-01-23
I think we still need some additional tests here around Select.