@roidelapluie 2020-01-23
yeah especially since I am still a beginner in TSDB :)