@brian-brazil 2020-01-23
There's also subtly different semantics around open/closed time ranges between PromQL and blocks.