@roidelapluie 2020-02-20
Rebased, squashed + added benchmarks to my git commit.