@brian-brazil 2020-02-21
:+1: 

I presume the benchmarks still look about the same after all the changes?