@roidelapluie 2020-02-21
Benchmarks are in both commit messages:

https://github.com/prometheus/prometheus/pull/6676/commits

You had -90%, I have -70%. There have been changes in TSDB since your original commit, not sure if related.