@roidelapluie 2020-02-21
mmh no benchmarks are both 90%

I guess I need to compare yours BenchmarkHeadSeries with BenchmarkQuerierSelect/SortedHead.

We have -70% in BenchmarkQuerierSelect/Head.