@brian-brazil 2020-02-21
No regression then, and we're now actually faster when all series are returned (which is surprising).